+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2025-07-14
+
+### Added
+
+- `Abgr8888` and `Rgba8888` are now `Copy, Clone, Debug, Eq, PartialEq, Ord, PartialOrd`
+- Optional feature `bytemuck` enables `Pod + Zeroable` for pixel representations
+
+### Changed
+
+- Removed `as_slice` and `as_slice_mut` in favor of (optional) `bytemuck`
+
 ## [0.2.0] - 2025-07-14
 
 ### Added